backend/requirements.txt
Django==3.2.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.1
pymongo==3.12.1
djongo==1.3.6
